Rename cbot package appropriately
